exp/requirements.txt
emoji==2.9.0
keras==2.10.0
Keras_Preprocessing==1.1.2
matplotlib==3.7.1
ml_collections==0.1.1
nltk==3.8.1
nni==3.0
numpy==1.24.3
openai==0.28.0
optuna==3.4.0
pandas==1.5.3
Pillow==9.4.0
Pillow==10.2.0
scikit_learn==1.2.2
scipy==1.10.1
skimage==0.0
tensorflow==2.10.0
torch==2.3.0.dev20240122+cu121
torchtext==0.15.2
torchvision==0.18.0.dev20240122+cu121
tqdm==4.66.1
transformers==4.36.2
